+++ lamindb/models/artifact.py
@@ -78,10 +78,8 @@ from ._relations import (
 from .base import (
     IsVersioned,
     LinkORM,
-    TracksRun,
-    TracksUpdates,
 )
-from .core import ParamValue, Storage, User
+from .core import Storage
 from .feature import Feature, FeatureValue
 from .has_parents import view_lineage
 from .record import (
@@ -91,7 +89,7 @@ from .record import (
     format_field_value,
     record_repr,
 )
-from .run import Run
+from .run import ParamValue, Run, TracksRun, TracksUpdates, User
 from .schema import Schema
 from .ulabel import ULabel
